Reworked TCPServer implementation
I pulled the TCP Server out of the Coordinator and implemented it within TCPServer in a more extensible manner. Now we can register different ProcessDelegates with the TCPServer to automatically direct requests against the server to the correct process. Please look at the Program.java startCoordinator method to see an example of this.